pyproject.toml
[tool.poetry]
name = "shirt-backend"
version = "0.1.0"
description = ""
authors = ["Hammad-Rehman <124163466+developer-hammad-rehman@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlmodel = "^0.0.19"
openai = "^1.31.0"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
